clippy & fmt

+++ yunet-utils/src/gpu/crop_batch.rs
@@ -112,7 +112,7 @@ impl GpuBatchCropper {
 
         let source_data = pack_rgba_pixels(rgba.as_raw());
         let source_buffer_size = (source_data.len() * std::mem::size_of::<u32>()) as u64;
-        let max_binding = self.context.limits().max_storage_buffer_binding_size as u64;
+        let max_binding = self.context.limits().max_storage_buffer_binding_size;
         anyhow::ensure!(
             source_buffer_size <= max_binding,
             "Source image too large for GPU batch cropping ({:.2} MB exceeds {:.2} MB WebGPU buffer binding limit). Image dimensions: {}x{}. Use CPU cropping for large images.",
